Move isEventLaterThisDay to utils/event
